% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cohortSim}
\alias{cohortSim}
\title{Simulation of a cohort}
\usage{
cohortSim(params, t_incr = 1)
}
\arguments{
\item{params}{List of parameters to use in function}

\item{t_incr}{Value to use a time increment (in years). Defaults to 1.}
}
\value{
A list
}
\description{
\code{cohortSim} simulates a single cohort
}
\examples{
data(tilapia)
res <- cohortSim(tilapia, t_incr=.1)
plot(pcap ~ Lt, res, t="l")
plot(Lt ~ t, res, t="l")
plot(Wt ~ t, res, t="l")

plot(Bt ~ t, res, t="l")
lines(SBt ~ t, res, col=2)

plot(Bt ~ Lt, res, t="l")
lines(SBt ~ Lt, res, col=2)

plot(Yt ~ t, res, t="l")

plot(Nt ~ t, res, t="l", log="y")
lines(Nt.noF ~ t, res, col=2, lty=2)
}

