% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_growth_soVB.R
\name{dt_growth_soVB}
\alias{dt_growth_soVB}
\title{Discrete seasonally oscillating von Bertalanffy growth function}
\usage{
dt_growth_soVB(Linf, K, ts, C, L1, t1, t2)
}
\arguments{
\item{Linf}{Infinite length}

\item{K}{growth constant}

\item{ts}{summer point. Time of year (between 0 and 1) when growth oscillation
cycle begins (sine wave term becomes positive)}

\item{C}{oscillation strength. Varies between 0 and 1.}

\item{L1}{Length at t1}

\item{t1}{time 1}

\item{t2}{time 2}
}
\description{
\code{dt_growth_soVB} calculates length at time 2, given 
length at time 1 according to the seasonally oscillationg 
von Bertalanffy growth function
}
\examples{
# growth parameters
Linf = 100
K = 0.5
ts = 0.5
C = 0.75

# create dataframe
df <- data.frame(t = seq(0,5,0.1), L=NaN)
df$L[1] <- 10 # intial size

# run iterative model
for(i in 2:nrow(df)){
  df$L[i] <- dt_growth_soVB(
    Linf, K, ts, C, 
    L1=df$L[i-1],
    t1=df$t[i-1],
    t2=df$t[i]
  )
}

# plot result
plot(L ~ t, df)
arrows(
  x0 = df$t[-nrow(df)], y0 = df$L[-nrow(df)],
  x1 = df$t[-1], y1 = df$L[-1],
  col = 8, length = 0.15
)

}
\references{
Somers, I. F. (1988). On a seasonally oscillating growth function. 
Fishbyte, 6(1), 8-11.
}

