% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fecW_LM.R
\name{fecW_LM}
\alias{fecW_LM}
\title{Fecundity and function of weight}
\usage{
fecW_LM(fec_a = 0, fec_b = 1, Wt = 500)
}
\arguments{
\item{fec_a}{intercept}

\item{fec_b}{slope}

\item{Wt}{individual weight (from \code{cohortSim} function)}
}
\value{
Number of eggs per individual
}
\description{
\code{fecW_LM} is a linear model describing fecundity (number of eggs) as a function
of individual weight
}
\examples{
data(tilapia)
res <- cohortSim(tilapia, t_incr=0.1)
plot(Neggst ~ Wt, res) # Number of eggs as a function of ind. weight
plot(res$t, res$Neggst*res$pmat*res$Nt) # total population fecundity by age

}
