% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_soVB.R
\name{growth_soVB}
\alias{growth_soVB}
\title{seasonally oscillating von Bertalanffy growth function}
\usage{
growth_soVB(Linf, K, t, t0, ts, C)
}
\arguments{
\item{Linf}{Infinite length}

\item{K}{growth constant}

\item{t}{age}

\item{t0}{(hypothetical) age at length zero}

\item{ts}{summer point. Time of year (between 0 and 1) when growth oscillation
cycle begins (sine wave term becomes positive)}

\item{C}{oscillation strength. Varies between 0 and 1.}
}
\description{
\code{growth_soVB} describes the growth as a function of age (t)
using the seasonally oscillating von Bertalanffy growth function (Somers 1988).
}
\examples{
t <- seq(0,5,0.1)
L <- growth_soVB(Linf=100, K=0.5, t=t, t0=-0.2, ts=0.5, C=0.75)
plot(t, L, t="l")

}
\references{
Somers, I. F. (1988). On a seasonally oscillating growth function. 
Fishbyte, 6(1), 8-11.
}

