% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_soVB2.R
\name{growth_soVB2}
\alias{growth_soVB2}
\title{seasonally oscillating von Bertalanffy growth function (using L0)}
\usage{
growth_soVB2(Linf, K, t, L0, ts, C)
}
\arguments{
\item{Linf}{Infinite length}

\item{K}{growth constant}

\item{t}{age}

\item{L0}{(hypothetical) length at time zero}

\item{ts}{summer point. Time of year (between 0 and 1) when growth oscillation
cycle begins (sine wave term becomes positive). Note that this definition 
differs from some interpretations of the model (see Somers 1998)}

\item{C}{oscillation strength. Varies between 0 and 1.}
}
\description{
\code{growth_soVB2} describes the growth as a function of age (t)
using the seasonally oscillating von Bertalanffy growth function (Somers 1988).
The function differs slightly from \code{\link[fishdynr]{growth_soVB}} 
by using \code{L0} instead of \code{t0} to describe the growth function origin.
}
\examples{
t <- seq(0,5,0.1)
L <- growth_soVB2(Linf=100, K=0.5, t=t, L0=10, ts=0.5, C=0.5)
plot(t, L, t="l", ylim=c(0,110), xaxs="i", yaxs="i")
points(0,10); abline(h=100, lty=2, col=8)

}

