% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth_VB.R
\name{growth_VB}
\alias{growth_VB}
\title{von Bertalanffy growth function}
\usage{
growth_VB(Linf, K, t, t0)
}
\arguments{
\item{Linf}{Infinite length}

\item{K}{growth constant}

\item{t}{age}

\item{t0}{(hypothetical) age at length zero}
}
\description{
\code{growth_VB} describes the growth as a function of age (t)
using the von Bertalanffy growth function
}
\examples{
t <- seq(0,5,0.1)
L <- growth_VB(Linf=100, K=0.5, t=t, t0=-0.2)
plot(t, L, t="l")
}

