% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/knife_edge.R
\name{knife_edge}
\alias{knife_edge}
\title{knife-edge selection}
\usage{
knife_edge(Lt, knife_edge_size)
}
\arguments{
\item{Lt}{body size}

\item{knife_edge_size}{knife edge size. Minimum size at capture.}
}
\description{
\code{knife_edge} describes knife-edge selection where
probability of capture 100\code{\%} after a minimum defined size. Provides
a rough estimate of trawl-type selectivity.
}
\examples{
data(tilapia)
tilapia$selectFun="knife_edge"
knife_edge_sizes <- c(20, 25, 30, 35)
for(i in seq(knife_edge_sizes)){
  tilapia$knife_edge_size <- knife_edge_sizes[i]
  res <- cohortSim(tilapia, t_incr=0.01)
  if(i == 1) plot(pcap ~ Lt, res, t="n")
  lines(pcap ~ Lt, res, col=i)
}
legend("topleft", legend=knife_edge_sizes,
  col=seq(knife_edge_sizes), lty=1,
  title="min. size", bty="n"
)
}

