% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfqGen.R
\name{lfqGen}
\alias{lfqGen}
\title{Generate length frequency data from a synthetic fish population}
\usage{
lfqGen(tincr = 1/12, K.mu = 0.5, K.cv = 0.1, Linf.mu = 80,
  Linf.cv = 0.1, ts = 0, C = 0.85, LWa = 0.01, LWb = 3, Lmat = 0.5 *
  Linf.mu, wmat = Lmat * 0.2, rmax = 10000, beta = 1, repro_wt = c(0, 0,
  1, 0, 0, 0, 0, 0, 0, 0, 0, 0), M = 0.7, harvest_rate = M, L50 = 0.25 *
  Linf.mu, wqs = L50 * 0.2, bin.size = 1, timemin = 0, timemax = 5,
  timemin.date = as.Date("1980-01-01"), N0 = 10000, fished_t = seq(0, 5,
  tincr), lfqFrac = 1)
}
\arguments{
\item{tincr}{time increment}

\item{K.mu}{mean K (growth parameter from von Bertalanffy growth function)}

\item{K.cv}{coefficient of variation on K
#@param L0.mu mean length at recruitment
#@param L0.cv coefficient of variation on L0}

\item{Linf.mu}{mean Linf (infinite length parameter from von Bertalanffy growth function)}

\item{Linf.cv}{coefficient of variation on Linf}

\item{ts}{summer point (range 0 to 1) (parameter from seasonally oscillating von Bertalanffy growth function)}

\item{C}{strength of seasonal oscillation (range 0 to 1) (parameter from seasonally oscillating von Bertalanffy growth function)}

\item{LWa}{length-weight relationship constant 'a' (W = a*L^b). Model assumed length in cm and weight in kg.}

\item{LWb}{length-weight relationship constant 'b' (W = a*L^b). Model assumed length in cm and weight in kg.}

\item{Lmat}{length at maturity (where 50\% of individuals are mature)}

\item{wmat}{width between 25\% and 75\% quantiles for Lmat}

\item{rmax}{parameter for Beverton-Holt stock recruitment relationship (see \code{\link[fishdynr]{srrBH}})}

\item{beta}{parameter for Beverton-Holt stock recruitment relationship (see \code{\link[fishdynr]{srrBH}})}

\item{repro_wt}{weight of reproduction (vector of monthly reproduction weight)}

\item{M}{natural mortality}

\item{harvest_rate}{Fishing mortality (i.e. 'F')}

\item{L50}{minimum length of capture (in cm). Where selectivity equals 0.5. Assumes logistic ogive typical of trawl net selectivity.}

\item{wqs}{width of selectivity ogive (in cm)}

\item{bin.size}{resulting bin size for length frequencies (in cm)}

\item{timemin}{time at start of simulation (in years). Typically set to zero.}

\item{timemax}{time at end of simulation (in years).}

\item{timemin.date}{date corresponding to timemin (of "Date" class)}

\item{N0}{starting number of individuals}

\item{fished_t}{times when stock is fished}

\item{lfqFrac}{fraction of fished stock that are sampled for length frequency data (default = 1).}

\item{tincr}{time increment for simulation (default = 1/12; i.e. 1 month)}
}
\value{
a list containing growth parameters and length frequency object
}
\description{
See (see \code{\link[fishdynr]{dt_growth_soVB}})
}
\examples{
\donttest{
res <- lfqGen()
names(res)

op <- par(mfcol=c(2,1), mar=c(4,4,1,1))
plot(N ~ dates, data=res$pop, t="l")
plot(B ~ dates, data=res$pop, t="l", ylab="B, SSB")
lines(SSB ~ dates, data=res$pop, t="l", lty=2)
par(op)

pal <- colorRampPalette(c("grey30",5,7,2), bias=2)
with(res$lfqbin, image(x=dates, y=midLengths, z=t(catch), col=pal(100)))
}


}

