% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticSelect.R
\name{logisticSelect}
\alias{logisticSelect}
\title{logistic selection}
\usage{
logisticSelect(Lt, L50, wqs)
}
\arguments{
\item{Lt}{body size}

\item{L50}{length at first capture (i.e. where prob. of capture equals 50\%)}

\item{wqs}{width between 25\% and 75\%  probability of capture}
}
\description{
\code{logisticSelect} describes knife-edge selection where 
probability of capture 100\code{\%} after a minimum defined size. Provides
a rough estimate of trawl-type selectivity.
}
\examples{
data(tilapia)
tilapia$selectFun="logisticSelect"
tilapia$L50 <- 20
WQS <- c(0, 3, 6)
COL <- rainbow(3)
for(i in seq(WQS)){
  tilapia$wqs <- WQS[i]
  res <- cohortSim(tilapia, t_incr=0.01)
  if(i == 1) plot(pcap ~ Lt, res, t="n")
  lines(pcap ~ Lt, res, col=COL[i])
}
legend("bottomright", legend=paste(WQS, c(" ('knife-edge')", "", ""), sep=""), 
  col=COL, lty=1, 
  title="Width between 25\% and 75\%", bty="n"
)
lines(c(0,tilapia$L50,tilapia$L50), c(0.5,0.5,0), col=1, lty=2)
text(20,.5,bquote("L"[50]), pos=4, col=1, font=1)

}

