% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtualPop.R
\name{virtualPop}
\alias{virtualPop}
\title{Virtual fish population}
\usage{
virtualPop(tincr = 1/12, K.mu = 0.5, K.cv = 0.1, Linf.mu = 80,
  Linf.cv = 0.1, ts = 0.5, C = 0.75, LWa = 0.01, LWb = 3, Lmat = 40,
  wmat = 8, rmax = 10000, beta = 1, repro_wt = c(0, 0, 0, 1, 0, 0, 0, 0,
  0, 0, 0, 0), M = 0.7, harvest_rate = M, L50 = 0.25 * Linf.mu,
  wqs = L50 * 0.2, bin.size = 1, timemin = 0, timemax = 10,
  timemin.date = as.Date("1980-01-01"), N0 = 10000, initializePop = TRUE,
  fished_t = seq(timemin, timemax, tincr), lfqFrac = 1,
  progressBar = TRUE)
}
\arguments{
\item{tincr}{time increment for simulation (default = 1/12; i.e. 1 month)}

\item{K.mu}{mean K (growth parameter from von Bertalanffy growth function)}

\item{K.cv}{coefficient of variation on K}

\item{Linf.mu}{mean Linf (infinite length parameter from von Bertalanffy growth function)}

\item{Linf.cv}{coefficient of variation on Linf}

\item{ts}{summer point (range 0 to 1) (parameter from seasonally oscillating von Bertalanffy growth function)}

\item{C}{strength of seasonal oscillation (range 0 to 1) (parameter from seasonally oscillating von Bertalanffy growth function)}

\item{LWa}{length-weight relationship constant 'a' (W = a*L^b). Model assumed length in cm and weight in kg.}

\item{LWb}{length-weight relationship constant 'b' (W = a*L^b). Model assumed length in cm and weight in kg.}

\item{Lmat}{length at maturity (where 50\% of individuals are mature)}

\item{wmat}{width between 25\% and 75\% quantiles for Lmat}

\item{rmax}{parameter for Beverton-Holt stock recruitment relationship (see \code{\link[fishdynr]{srrBH}})}

\item{beta}{parameter for Beverton-Holt stock recruitment relationship (see \code{\link[fishdynr]{srrBH}})}

\item{repro_wt}{weight of reproduction (vector of monthly reproduction weight)}

\item{M}{natural mortality}

\item{harvest_rate}{Fishing mortality (i.e. 'F')}

\item{L50}{minimum length of capture (in cm). Where selectivity equals 0.5. Assumes logistic ogive typical of trawl net selectivity.}

\item{wqs}{width of selectivity ogive (in cm)}

\item{bin.size}{resulting bin size for length frequencies (in cm)}

\item{timemin}{time at start of simulation (in years). Typically set to zero.}

\item{timemax}{time at end of simulation (in years).}

\item{timemin.date}{date corresponding to timemin (of "Date" class)}

\item{N0}{starting number of individuals}

\item{initializePop}{logical. Should initial population be approximated based on SRR equilibrium}

\item{fished_t}{times when stock is fished}

\item{lfqFrac}{fraction of fished stock that are sampled for length frequency data (default = 0.1).}

\item{progressBar}{Logical. Should progress bar be shown in console (Default=TRUE)}
}
\value{
a list containing growth parameters and length frequency object
}
\description{
See \code{\link[fishdynr]{dt_growth_soVB}} for information on growth function.
The model creates variation in growth based on a mean phi prime value for the population,
which describes relationship between individual Linf and K values. See Vakily (1992) 
for more details.
}
\examples{

set.seed(1)
res <- virtualPop(initializePop = T, N0=10000)
names(res)

op <- par(mfcol=c(2,1), mar=c(4,4,1,1))
plot(N ~ dates, data=res$pop, t="l")
plot(B ~ dates, data=res$pop, t="l", ylab="B, SSB")
lines(SSB ~ dates, data=res$pop, t="l", lty=2)
par(op)

pal <- colorRampPalette(c("grey30",5,7,2), bias=2)
with(res$lfqbin, image(x=dates, y=midLengths, z=t(catch), col=pal(100)))

### biased results with single monthly sample
inds <- res$inds[[1]]
plot(mat ~ L, data = inds, pch=".", cex=5, col=rgb(0,0,0,0.05))
fit <- glm(mat ~ L, data = inds, family = binomial(link = "logit"))
summary(fit)

newdat <- data.frame(L = seq(min(inds$L), max(inds$L), length.out=100))
newdat$pmat <- pmat_w(newdat$L, Lmat = 40, wmat=40*0.2)
pred <- predict(fit, newdata=newdat, se.fit=TRUE)

# Combine the hypothetical data and predicted values
newdat <- cbind(newdat, pred)
# Calculate confidence intervals
std <- qnorm(0.95 / 2 + 0.5)
newdat$ymin <- fit$family$linkinv(newdat$fit - std * newdat$se.fit)
newdat$ymax <- fit$family$linkinv(newdat$fit + std * newdat$se.fit)
newdat$fit <- fit$family$linkinv(newdat$fit)  # Rescale to 0-1

plot(mat ~ L, data = inds, pch=".", cex=5, col=rgb(0,0,0,0.05))
lines(pmat ~ L, newdat, col=8, lty=3)
polygon(
  x = c(newdat$L, rev(newdat$L)),
  y = c(newdat$ymax, rev(newdat$ymin)),
  col = adjustcolor(2, alpha.f = 0.3),
  border = adjustcolor(2, alpha.f = 0.3)
)
lines(fit ~ L, newdat, col=2)

lrPerc <- function(alpha, beta, p) (log(p/(1-p))-alpha)/beta
( L50 <- lrPerc(alpha=coef(fit)[1], beta=coef(fit)[2], p=0.5) )
lines(x=c(L50,L50,0), y=c(-100,0.5,0.5), lty=2, col=2)
text(x=L50, y=0.5, labels = paste0("L50 = ", round(L50,2)), pos=4, col=2 )



### all samples combined
inds <- do.call("rbind", res$inds)
plot(mat ~ L, data = inds, pch=".", cex=5, col=rgb(0,0,0,0.05))
fit <- glm(mat ~ L, data = inds, family = binomial(link = "logit"))
summary(fit)

newdat <- data.frame(L = seq(min(inds$L), max(inds$L), length.out=100))
newdat$pmat <- pmat_w(newdat$L, Lmat = 40, wmat=40*0.2)
pred <- predict(fit, newdata=newdat, se.fit=TRUE)
# Combine the hypothetical data and predicted values
newdat <- cbind(newdat, pred)
# Calculate confidence intervals
std <- qnorm(0.95 / 2 + 0.5)
newdat$ymin <- fit$family$linkinv(newdat$fit - std * newdat$se.fit)
newdat$ymax <- fit$family$linkinv(newdat$fit + std * newdat$se.fit)
newdat$fit <- fit$family$linkinv(newdat$fit)  # Rescale to 0-1

plot(mat ~ L, data = inds, pch=".", cex=5, col=rgb(0,0,0,0.05))
lines(pmat ~ L, newdat, col=8, lty=3)
polygon(
  x = c(newdat$L, rev(newdat$L)),
  y = c(newdat$ymax, rev(newdat$ymin)),
  col = adjustcolor(2, alpha.f = 0.3),
  border = adjustcolor(2, alpha.f = 0.3)
)
lines(fit ~ L, newdat, col=2)

lrPerc <- function(alpha, beta, p) (log(p/(1-p))-alpha)/beta
( L50 <- lrPerc(alpha=coef(fit)[1], beta=coef(fit)[2], p=0.5) )
lines(x=c(L50,L50,0), y=c(-100,0.5,0.5), lty=2, col=2)
text(x=L50, y=0.5, labels = paste0("L50 = ", round(L50,2)), pos=4, col=2 )





}
\references{
Vakily, J.M., 1992. Determination and comparison of bivalve growth, 
with emphasis on Thailand and other tropical areas. WorldFish.

Munro, J.L., Pauly, D., 1983. A simple method for comparing the growth 
of fishes and invertebrates. Fishbyte 1, 5-6.

Pauly, D., Munro, J., 1984. Once more on the comparison of growth 
in fish and invertebrates. Fishbyte (Philippines).
}
