% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cohortSim.R
\name{cohortSim}
\alias{cohortSim}
\title{Simulation of a cohort}
\usage{
cohortSim(params, t_incr = 1)
}
\arguments{
\item{params}{List of parameters to use in function}

\item{t_incr}{Value to use a time increment (in years). Defaults to 1.}
}
\value{
A list
}
\description{
\code{cohortSim} simulates a single cohort
}
\details{
\code{params} list should contain the following parameters:
\itemize{
  \item \code{species} Species name
  \item \code{growthFun} Name of growth function (e.g. "growth_VB" is the von
  Bertalanffy growth function)
  \item \code{amax} Maximum age
  \item \code{LWa} Length-weight relationship parameter a (weight~a*length^b)
  \item \code{LWb} Length-weight relationship parameter b (weight~a*length^b)
  \item \code{M} Natural mortality
  \item \code{F} Fishing mortality
  \item \code{N0} Number of individuals at time 0
  \item \code{matFun} Name of maturity function (e.g. "pmat_w" is a logistic
  function that includes width, w, of quantiles)
  \item \code{selectFun} Function to use for gear selection. Determines lengths
  vulnerable to fishing mortality (e.g. "gillnet" and "knife_edge" functions).
  \item \code{...} Other parameters for growth, maturity, and selectivity functions.
}
}
\examples{
data(tilapia)
res <- cohortSim(tilapia, t_incr=.1)
plot(pcap ~ Lt, res, t="l")
plot(Lt ~ t, res, t="l")
plot(Wt ~ t, res, t="l")

plot(Bt ~ t, res, t="l")
lines(SBt ~ t, res, col=2)

plot(Bt ~ Lt, res, t="l")
lines(SBt ~ Lt, res, col=2)

plot(Yt ~ t, res, t="l")

plot(Nt ~ t, res, t="l", log="y")
lines(Nt.noF ~ t, res, col=2, lty=2)
}

