% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/optim.stockSim.R
\name{optim.stockSim}
\alias{optim.stockSim}
\title{Optimization of fishing mortality}
\usage{
optim.stockSim(Ft = 0, params, nyears = 100, envKt = 1, envSt = 1)
}
\arguments{
\item{Ft}{time series vector for fishing mortality. If a single value, then
the function assumes a constant fishing mortality for the entire simulation
(default=0)}

\item{params}{List of parameters to for the population. Applies a single
cohort simulation for the initial population state using
\code{\link[fishdynr]{cohortSim}} function. See
\code{\link[fishdynr]{stockSim}} for details.}

\item{nyears}{number of years in the simulation}

\item{envKt}{time series vector for environmental effects to maximum
recruitment (e.g. \code{rmax} in \code{\link[fishdynr]{srrBH}})
(default=1).}

\item{envSt}{time series vector for environmental effects to half maximum
  recruitment parameter (e.g. \code{beta} in \code{\link[fishdynr]{srrBH}})
  (default=1).}
}
\description{
Sets up an exploration of optimum fishing policy using the function
\code{\link[fishdynr]{stockSim}}
(i.e. optimal time series of fishing mortalities, \code{Ft}, in order to maximize
time series yield, \code{Yt}). The routine is described in further detail in
Walters and Martell (2004).
}
\examples{
\donttest{
data(tilapia)
params <- tilapia
params$N0 <- 1e8
params$knife_edge_size <- 20
nyears <- 100
Ft <- rep(0.5, nyears)
envKt <- rep(1, nyears); envKt[30:50] <- 0.5
envSt <- runif(nyears, min=0.5, max=1.5)

# Optimization of Ft (will take some time to reach cost function mimimum)
out <- optim(
  par = Ft,  # initial guess
  fn = optim.stockSim,
  params = params,
  nyears = nyears,
  envKt = envKt,
  envSt = envSt,
  method = "L-BFGS-B",
  lower = 0,
  upper = 3,
  control = list(fnscale=-1, trace=4)
)

# optimum Ft series
plot(out$par, t="l")

# optimum Yt series
tmp <- stockSim(params, nyears=nyears, Ft=out$par, envKt=envKt, envSt=envSt)
plot(Yt ~ t, tmp, t="l")
sum(tmp$Yt/1e6, na.rm=TRUE)

# optimum yield versus stock biomass
plot(Yt ~ Bt, tmp)
fit <- lm(Yt ~ Bt, tmp)
abline(fit)
summary(fit)
plot(Bt ~ t, tmp, t="l", ylim=c(0, max(tmp$Bt)))
lines(Yt ~ t, tmp, col=2)
}
}
\references{
Walters, C. J., Martell, S. J., 2004.
Fisheries ecology and management. Princeton University Press.
}

