% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{stockSim}
\alias{stockSim}
\title{Simulation of a stock}
\usage{
stockSim(params, nyears = 100, Ft = 0, envKt = 1, envSt = 1)
}
\arguments{
\item{params}{List of parameters to for the population. Applies a single cohort
simulation for the initial population state using \code{\link[fishdynr]{cohortSim}}
function}

\item{nyears}{number of years in the simulation}

\item{Ft}{time series vector for fishing mortality. If a single value, then the function
assumes a constant fishing mortality for the entire simulation (default=0)}

\item{envKt}{time series vector for environmental effects to maximum recruitment
(e.g. \code{rmax} in \code{\link[fishdynr]{srrBH}}) (default=1).}

\item{envSt}{time series vector for environmental effects to half maximum recruitment
parameter (e.g. \code{beta} in \code{\link[fishdynr]{srrBH}}) (default=1).}
}
\value{
A list
\itemize{
  \item \code{Btc} matrix. Stock biomass by time (rows) and cohort (columns).
  \item \code{Ytc} matrix. Fishery yield by time (rows) and cohort (columns).
  \item \code{Bt} vector. Stock biomass by time.
  \item \code{Yt} vector. Fishery yield by time.
  \item \code{Nt} vector. Stock size (in numbers) by time.
  \item \code{Ct} vector. Fishery catch (in numbers) by time.
}
}
\description{
\code{stockSim} simulates a stock using variable fishing mortality
and stock recruitment relationship
}
\details{
\code{params} list should contain the following parameters:
\itemize{
  \item \code{species} Species name
  \item \code{growthFun} Name of growth function (e.g. "growth_VB" is the von
  Bertalanffy growth function)
  \item \code{amax} Maximum age
  \item \code{LWa} Length-weight relationship parameter a (weight~a*length^b)
  \item \code{LWb} Length-weight relationship parameter b (weight~a*length^b)
  \item \code{M} Natural mortality
  \item \code{F} Fishing mortality
  \item \code{N0} Number of individuals at time 0
  \item \code{matFun} Name of maturity function (e.g. "pmat_w" is a logistic
  function that includes width, w, of quantiles)
  \item \code{selectFun} Function to use for gear selection. Determines lengths
  vulnerable to fishing mortality (e.g. "gillnet" and "knife_edge" functions).
  \item \code{srrFun} Stock-recruitment relationship function (e.g. "srrBH").
  \item \code{fec} Number of eggs produced per weight [g] of mature female (For use in
  srrFun).
  \item \code{...} Other parameters for growth, maturity, and selectivity functions.
}
For fitting an optimal time series of fishing mortalities, \code{Ft}, see
\code{\link[fishdynr]{optim.stockSim}} (Walters and Martell, 2004).
}
\examples{
data(tilapia)
params <- tilapia
params$knife_edge_size <- 20
params$rmax <- 1e6
params$beta <- 1e9
params$fec <- 100
params$N0 <- 1e5
nyears <- 100
Ft <- rep(0.5, nyears); Ft[20:40] <- 1
envKt <- rep(1, nyears); envKt[50:100] <- 0.5
envSt <- runif(nyears, min=0.8, max=1.2)
tmp <- stockSim(Ft=Ft, params=params, nyears=100, envKt=envKt, envSt=envSt)
plot(tmp$Bt, t="l")
plot(tmp$Yt, t="l")
sum(tmp$Yt/1e6, na.rm=TRUE)
}
\references{
Walters, C. J., Martell, S. J., 2004.
Fisheries ecology and management. Princeton University Press.
}

