% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{stockSim}
\alias{stockSim}
\title{Simulation of a stock}
\usage{
stockSim(params, nyears = 100, Ft = 0, envKt = 1, envSt = 1)
}
\arguments{
\item{params}{List of parameters to for the population. Applies a single cohort
simulation for the initial population state using \code{\link[fishdynr]{cohortSim}}
function}

\item{nyears}{number of years in the simulation}

\item{Ft}{time series vector for fishing mortality. If a single value, then the function
assumes a constant fishing mortality for the entire simulation (default=0)}

\item{envKt}{time series vector for environmental effects to maximum recruitment
(e.g. \code{rmax} in \code{\link[fishdynr]{srrBH}}) (default=1).}

\item{envSt}{time series vector for environmental effects to half maximum recruitment
parameter (e.g. \code{beta} in \code{\link[fishdynr]{srrBH}}) (default=1).}
}
\value{
A list
}
\description{
\code{stockSim} simulates a stock using variable fishing mortality
and stock recruitment relationship
}
\examples{
data(tilapia)
params <- tilapia
params$knife_edge_size <- 20
params$rmax <- 1e6
params$beta <- 1e9
params$fec <- 100
params$N0 <- 1e5
nyears <- 100
Ft <- rep(0.5, nyears); Ft[20:40] <- 1
envKt <- rep(1, nyears); envKt[50:100] <- 0.5
envSt <- runif(nyears, min=0.8, max=1.2)
tmp <- stockSim(params, nyears=100, Ft=Ft, envKt=envKt, envSt=envSt)
plot(tmp$Bt, t="l")
plot(tmp$Yt, t="l")
sum(tmp$Yt/1e6, na.rm=T)
}

